fix(game): Garante que as missões sejam resetadas corretamente
A função `restartGame` estava resetando as missões para um estado fixo ("hardcoded"), ignorando a configuração inicial definida em `config.js`. Isso criava duas fontes de verdade e impedia que alterações na configuração inicial das missões fossem refletidas após o reinício do jogo.

Esta correção resolve o problema da seguinte forma:
1. Uma cópia profunda da configuração inicial das missões é armazenada no início do jogo.
2. A função `restartGame` agora usa essa cópia para restaurar o estado das missões, garantindo que `config.js` seja a única fonte de verdade.
3. Um novo teste foi adicionado para verificar se as missões são resetadas para seu estado original, prevenindo regressões futuras.
4. O ambiente de teste foi aprimorado com um arquivo `jest.setup.js` para mockar o DOM e o áudio, tornando os testes mais robustos.
