+++ jest.config.cjs
@@ -1,3 +1,4 @@
 module.exports = {
   testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['./jest.setup.js'],
 };
